requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
packaging==24.2
propcache==0.2.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.2.3
yarl==1.18.3